Implement Job Announcement feature with CRUD operations, including controller, service, repository, and DTO; enhance frontend handling and validation

+++ backend/src/main/java/com/nextjob/repositories/ContractTypeRepository.java
@@ -3,4 +3,4 @@ package com.nextjob.repositories;
 import com.nextjob.entities.ContractType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractTypeRepository extends JpaRepository<ContractType, Long> {}
+public interface ContractTypeRepository extends JpaRepository<ContractType, Integer> {}
