+++ backend/src/main/java/com/nextjob/repositories/PaymentTypeRepository.java
@@ -3,4 +3,4 @@ package com.nextjob.repositories;
 import com.nextjob.entities.PaymentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentTypeRepository extends JpaRepository<PaymentType, Long> {}
+public interface PaymentTypeRepository extends JpaRepository<PaymentType, Integer> {}
