+++ backend/src/main/java/com/nextjob/repositories/WorkModeRepository.java
@@ -3,4 +3,4 @@ package com.nextjob.repositories;
 import com.nextjob.entities.WorkMode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkModeRepository extends JpaRepository<WorkMode, Long> {}
+public interface WorkModeRepository extends JpaRepository<WorkMode, Integer> {}
